tests/python/requirements.txt
pytest>=7.0
azure-storage-blob>=12.19
